social_media_api/requirements.txt
asgiref==3.11.0
Django==5.2.10
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysql-connector-python==9.5.0
mysqlclient==2.2.7
pillow==12.1.1
PyJWT==2.11.0
PyMySQL==1.1.2
sqlparse==0.5.5
